requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
black==25.1.0
boto3==1.40.14
botocore==1.40.14
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
coverage==7.10.4
cryptography==45.0.6
distlib==0.4.0
filelock==3.19.1
flake8==7.3.0
identify==2.6.13
iniconfig==2.1.0
invoke==2.2.0
jmespath==1.0.1
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
paramiko==4.0.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
psutil==7.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyNaCl==1.5.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
s3transfer==0.13.1
six==1.17.0
structlog==25.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.34.0
